Move CLI support helpers out of main
